+++ services/liaradb/recovery/pagepool/page_pool.go
@@ -10,7 +10,7 @@ type PagePool struct {
 	pool sync.Pool
 }
 
-func New(size int16, slotHeaderSize int16) *PagePool {
+func New(size int, slotHeaderSize int) *PagePool {
 	return &PagePool{
 		pool: sync.Pool{New: func() any {
 			return logpage.New(size, slotHeaderSize)
